requirements.txt
# requirements.txt - FINAL FIX VERSION
# ✅ Addresses NumPy 2.x incompatibility with ultralytics

# Core Streamlit
streamlit>=1.28.0

# Interactive components  
streamlit-image-coordinates>=0.1.6
plotly>=5.15.0

# 🎯 EXACT MATCH: Your working local ML/CV versions
torch==2.6.0
torchvision==0.21.0
ultralytics==8.3.92

# 🚨 CRITICAL FIX: NumPy 1.x for ultralytics compatibility
# This is the ROOT CAUSE of your crashes!
numpy<2.0

# Alternative specific version if above doesn't work:
# numpy==1.24.3

# 🚨 CRITICAL: Headless OpenCV for Streamlit Cloud
opencv-python-headless==4.11.0.86

# 🎯 Compatible image/data processing versions
Pillow==11.1.0
scipy==1.15.2
pandas==2.2.3

# Utilities - matching your local versions
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
PyYAML==6.0.2
tqdm
psutil==7.0.0
requests>=2.25.0
supabase>=2.0.0